Added the HTML Link redirect tracking page
This page records the token parameter in the URL into local storage (should be substituted for a database or external file when used) and redirects to the skillsbuild page. The intended use case is to send a link to this page in an email with a specific and unique token so that the specific user that clicked on it can be determined.